fix: cast samples to i64 for rusqlite 0.39 ToSql compatibility
usize was removed from rusqlite's ToSql impls in 0.39; cast to i64 at the params! call site.

+++ src/storage/history.rs
@@ -429,7 +429,7 @@ impl HistoryDb {
 
         let mut durations = Vec::new();
 
-        let rows = stmt.query_map(params![samples], |row| {
+        let rows = stmt.query_map(params![samples as i64], |row| {
             let submitted_str: String = row.get(0)?;
             let completed_str: String = row.get(1)?;
